// frontend/src/components/ui/FormInput.d.ts
import { type InputHTMLAttributes } from 'react';
export interface FormInputProps extends InputHTMLAttributes<HTMLInputElement> {
    label?: string;
    error?: string;
    hint?: string;
}
export declare const FormInput: import("react").ForwardRefExoticComponent<FormInputProps & import("react").RefAttributes<HTMLInputElement>>;
//# sourceMappingURL=FormInput.d.ts.map